+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [🔧] 键盘高度没有获取到的时候, `stableHeight`使用null, 而不是0
 - [🚀] `ImageView`增加`cacheKey`参数透传
 - [🚀] `DecoratedText`增加`gradient`参数, 实现渐变色渲染
+- [🚀] `SelectableListX`增加`selectedCount`参数, 获取选中元素的个数
 
 ## 0.52.1
 - [⚠️] 将`Constellation`枚举重命名为`Zodiac`, `constellation`getter重命名为`zodiac`, 旧术语保留deprecated别名;